feat!: rename install -> create-alias, fetch -> download
