+++ crates/stellar-registry-cli/src/commands/download.rs
